Move RGW pod existence checks to their own test
